Hide the help bubble if no help available for field.
Problem: When in a form only some of the components have context help,
and followFocus is true, something strange happens. When the focus moves
to a field without context help, the context help of the previous
focussed field is also showed for the newly focussed field. In other
words, if only 1 field has context help, this is also shown for al the
other fields and the panel.

This fix, contributed by Mark Vlems, hides the help bubble if no help
text is available for the field.